[stats]: moved Histogram model from nduser models to stats models
