+++ django/stats/models.py
@@ -14,4 +14,21 @@
 
 from django.db import models
 
-# Create your models here.
+from nduser.models import Channel
+
+class Histogram (models.Model):
+  """ Stores a histogram in npz format """
+  channel = models.ForeignKey(Channel)
+  histogram = models.BinaryField(max_length=4096, null=True)
+  bins = models.BinaryField(max_length=4096, null=True)
+  REGION_CHOICES = (
+    (0, 'Entire Dataset'),
+    (1, 'ROI (rectangle)'), # ROI is a rectangle
+    (2, 'RAMON')
+  )
+  region = models.IntegerField(choices=REGION_CHOICES, default=0)
+  roi = models.TextField(blank=True) # stores JSON representation of rectangular ROI
+
+  class Meta:
+    db_table = u"histogram"
+    managed = True
